requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
config==0.5.0.post0
Cython==0.29.21
idna==2.10
pip-tools==5.3.1
pyTelegramBotAPI==3.7.2
requests==2.24.0
six==1.15.0
urllib3==1.25.10
vedis==0.7.1
rusyllab @ git+https://github.com/Koziev/rusyllab@c5d0ea405e9cc8233bd65503c26c0d2ad8b5cb57